Projeto Java

+++ src/ProjetoJava/ProjetoDeParcelaDoCarro.java
@@ -2,8 +2,8 @@ package ProjetoJava;
 
 public class ProjetoDeParcelaDoCarro {
     public static void main(String[] args) {
-        int valorDaParcela;
-        int valorDoCarro = 40000;
+        float valorDaParcela;
+        float valorDoCarro = 40000;
         for (int parcela = 1; parcela <= valorDoCarro; parcela++){
             valorDaParcela =   valorDoCarro / parcela;
             if(valorDaParcela >= 0){
